requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pillow==9.5.0
torch==2.0.0
torchvision==0.15.1